Refine the output of `EXPLAIN WITH(filter_pushdown)`
Previously, we'd consider the output of a function "pushdownable" if any
of its arguments could be predicted based on the range info we track.
This ends up being too generous: for example, `col + hash(col)` is not
predictable even if the left-hand side is.

Getting this right requires distinguishing between expressions that
don't depend on the column ranges because they're constants, and
expressions that don't depend on the column ranges because they're
complicated functions... so we have to change up the summary type here
as well.