storage: handle transient errors during decoding
Change the function signatures of all methods that can fail transiently
to be a double `Result` where the outer `Result` represents whether a
transient error happened and the inner `Result` is the definite result
of decoding the bytes.

This separates things that should be permanently recorded in the shard
(the inner errors) from things that should simply restart the dataflow
(outer errors). At the same time it makes it easy to propagate a
transient error upwards (just using the `?` operator) and hard to
accidentally put a transient error in the shard (since you have to
manually deconstruct the error and put it there).

This bubbling up of transient errors terminates in the future passed to
`build_fallible` which was integrated in #20150 and will restart the
ingestion dataflow if a transient error reaches that point.

Fixes #6185

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>