adapter: fork old implementations for `SELECT` paths
Fork the old implementations of

- `Coordinator::sequence_peek*`, and
- `Coordinator::explain_query_optimizer_pipeline`,

by duplicating code and dispatching between the old and the new code
path based on the value of the `enable_unified_optimizer_api` feature
flag. At the moment the two code paths are still identical, but this
will change with the next commit.

See the rollout proposal in MaterializeInc#20569 for details.