test: make tracing test more resilient to increased debug logging
This test can timeout on a PR if it adds a bunch of extra debug logging.
We intend for debug logging to be something that can be turned on
selectively, so this is a bug, not a feature. Make the test resilient to
this sort of thing by altering the less-selective filters to keep all
the mz crates at the same level and change things for a fictional "foo"
crate.