SQL: Make dangerous functions more granular
Previously, all functions within the `mz_internal` schema were declared
dangerous, and we disallowed users from executing any of them. This was
too broad and prevented users from executing useful, but not stabilized
functions, such as `is_rbac_enabled`.

This commit updates functions so that each one is annotated with a
dangerous flag that indicates if it's safe for a user to execute them.