persist-txn: hook enable_persist_txn_tables up to Launch Darkly
Only takes effect on restart. Any changes will also cause clusterd
processes to restart.

The usual Launch Darkly system vars hookup is not available early enough
in the boot process to use directly, so we mirror it to an entry in the
catalog stash "config" collection and read that. Leave the environmentd
cli flag around as an override to make testing easier.