Reuse allocation in FromRowByTypes
Previously, the implementation for Row would allocate a new row and extend
it with the provided iterator. This can be problematic because we don't
size the underlying buffer correctly, causing doubling allocations each
time we run out of space.

This change addresses the problem by reusing an underlying allocation as it
is done elsewhere, and in the process switching the trait to operator on
self instead of free-standing functions. We also add a try_from_datum_iter
variant that permits passing an iterator of results, saving an
intermediate packing/unpacking of rows.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>