sql: Correct timezone parsing to account for POSIX
PostgreSQL uses different timezone specification for parsing timezone
offsets depending on the context or the parsing. When parsing a type,
such as `TIMESTAMPTZ`, PostgreSQL uses ISO 8601 to parse the timezone.
However, for all other timezone parsing, such as the `timezone`
function or `SET TIMEZONE`, PostgreSQL uses the POSIX standard to
parse the timezone.

The PostgreSQL docs are fairly confusing on this matter, so there isn't
a single doc to link. As of version 16, the relevant information is
spread out across the following docs:

  - https://www.postgresql.org/docs/16/functions-datetime.html#FUNCTIONS-DATETIME-ZONECONVERT
  - https://www.postgresql.org/docs/16/datatype-datetime.html#DATATYPE-DATETIME-INPUT-TIMES
  - https://www.postgresql.org/docs/16/datatype-datetime.html#DATATYPE-TIMEZONES
  - https://www.postgresql.org/docs/16/datetime-posix-timezone-specs.html
  - https://www.postgresql.org/docs/16/sql-set.html

Previously, Materialize always used ISO 8601 to parse timezone offsets.
This commit updates the timezone parsing logic to use the correct
timezone in the correct context.

Fixes #21999