testdrive: add retry modifiers and RTR commands
This commit adds two features in testdrive that work together to
eliminate the vast majority of query retries in testdrive.

The simplest of the two changes is making all queries oneshot unless
specified otherwise. Specifying the retry behavior is done by adding the
`retry` or `noretry` keyword in the  modifier brackets. So `>[retry]
SELECT 1` specifies that the query should be retried if the result
doesn't match.

Retries are only allowed to be specified for SELECT queries and are
required if the query references any objects that are known to be
eventually consistent. The set of eventually consistent objects consists
of all the builtin sources and any builtin views that transitively
depend on any builtin source. This set is dynamically discovered by
testdrive on startup by introspecting Materialize using the `mz_objects`
and `mz_object_dependencies` tables.

The second change is a "poor man's" implementation of real time recency
that allows one to linearize a source on-demand. Two new commands are
introduced: `kafka-await-ingestion` and `postgres-await-ingestion` that
get parameterized by a source name and a topic/postgres connection
string. These commands should be used after a write to the upstream
system and before a read in Materialize that expects to see the write.
Here is a concrete example:

```
> CREATE SOURCE foo FROM KAFKA TOPIC foo

# Write to the upstream system
$ kafka-ingest topic=foo
some data

# Linearize the write
$ kafka-await-ingestion source=foo topic=foo

# Guarnateed to see the write with no retries
> SELECT * FROM foo
some data
```

These commands work by querying the respective upstream system (kafka or
postgres) for their current high watermark and comparing it with the
frontier recorded in the progress collection. Both the fetched high
watermark and the contents of the progress collection represent a
frontier and the commands will repeatedly `SELECT * FROM
source_progress` until that frontier is beyond the obtained high
watermark. When that happens we are free to run any `SELECT` queries on
the source and expect them to return results that include everything
up to the high watermark. This is because the progress collection
contains at timestamp `t` the data that the source collection will have
at time `t` and also because in strict serializable mode if we observe
the desired frontier at some `t_1` then all future queries will run at
timestamps `t_q >= t_1`.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>