stash-debug: Correctly retry opening the catalog
Add up to 30 seconds' worth of retries for the upgrade-check tool to
open the catalog, enabling us to check in spite of transaction errors
that might occur when running against a live environment.

The existing code would retry when opening the connection, but once the
connection was open would assume everything succeeded; this retries the
whole process.