adapter: bootstrap: create all storage collections upfront
This commit changes the bootstrap logic to create all collections, not
just tables and system sources, upfront. This has the benefit of being
more efficient.

This change is also a likely precondition for fixing a current
bug with selecting index as-of frontiers, as it allows the as-of
selection logic to query the storage controller about the frontiers of
dependent collections, which would not have been registered with the
storage controller otherwise.

The change in the storage controller is required to ensure that the
`create_collections` call doesn't fail because the IDs of status and
progress collections are missing from `durable_metadata` during
enrichment.