sql: add timezone_offset function
The timezone_offset function describes a time zone's offset from UTC at
a given moment in time. See the user-facing docs within the patch for
details.

This function will primarily be used in a future commit to implement the
`pg_timezone_names` and `pg_timezone_abbrevs` views, but it is exposed
publicly as it may be of use to end users too.