sql: support all Kafka security protocols
Through an accident of SQL parsing, we only supported the `SASL_SSL` and
`SSL` Kafka security protocols, and the latter required TLS client
authentication.

This commit teaches Materialize to also support:

  * `SECURITY PROTOCOL = PLAINTEXT`
  * `SECURITY PROTOCOL = SASL_PLAINTEXT`
  * `SECURITY PROTOCOL = SSL` without client authentication

This is, at last, the full matrix of Kafka security protocols.

Now that we support TLS without client authentication (previously the
only way to get SSL encryption was to use SASL or specify a client
certificate), we can have a more secure default. Now, SSL is always used
by default (i.e., the security protocol defaults to either SSL or
SASL_SSL depending on the presence of SASL options), and users must
explicitly downgrade to PLAINTEXT or SASL_PLAINTEXT. This is technically
a backwards incompatible change, but we add a migration to ensure that
all existing Kafka connections have their security protocols set
explicitly using the old rules, so that only new sinks are subject to
the new behavior.

Fix #12401.
Fix #19039.