adapter: move timestamp resolution after optimization
This commit moves timestamp resolution for dataflows implementing
indexes and materialized views after the full optimization.

The rationale for why this is sound is that the plans of standing
dataflows should never be affected by the selected as-of, so as-of
selection can happen whenever.

The advantage of moving as-of selection to the end is that this allows
us to pull optimization before as-of selection during bootstrap, so
as-of selection can make use of index dependency information discovered
during optimization.