persist: actually hook up Batch deletion to blob deletion
This was "temporarily" commented out ~8~ 17 months ago (apparently)
because of a nemesis failure and was never re-enabled.

We now not only have s3 object versioning, but also a persistcli
`restore-blob` command, which should resolve any of these if they do pop
up, making this much less stressful than the last time we tried this in
Jan. (I still plan to let this back in staging for a while.)