persist: replace should_fetch_part semantics with should_filter_part
The semantics of should_fetch_part require it to turn "I don't know"
into "true", which in practice has seemed to be a touch confusing. This
was confirmed in the code review for #22838 where Joe encountered
pushdown for the first time. Instead, flip the semantics to
should_filter_part, where "I don't know" (e.g. a `None`) gets mapped to
"false, which we hope will be easier to reason about when reading code.