adapter: never remove timeline state
This is motivated by the addition of an external, distributed timestamp
oracle which makes it harder to remove timeline state atomically with
the DDL/catalog/stash changes that cause said removal.

Additionally, it seems incorrect that a timelines' timestamp state can
vanish and then be re-created in the future, potentially at a lower
timestamp, which would violate linearizability.