optimizer: use the correct index key in `OptimizePeek`
It turns out that the `RelationType` (and more specifically its `keys`
property) might change between the end of the local MIR optimization
phase and the end of the global MIR optimization phase.

Consequently, the key that is attached to the exported index might be
different than the key of the fully optimized dataflow. This can cause
messed up plan construction because we currently use the former when
constructing the `IndexDesc` and the latter when figuring out the
permutation and thinning required to read from it the

```rust
permutation_for_arrangement(&key, arity);
```

call when constructing a `PeekPlan::SlowPath` in `plan_peek`. This
commit fixes the issue by passing the original type as a separate field
in the `GlobalMirPlan` result for `OptimizePeek`.

Fixes #23015.