storage: metric for count of in-memory remap bindings
During incident 75 we observed the CPU spending all its time in
reclocking. At the time the only theory we could come up with was that
the in-memory trace of the remap shard had not been compacted leading to
the slowdown of each reclock operation.

We couldn't confirm this was the case by neither observing the machines
during the incident nor by benchmarking, whose results were
inconclusive.

This PR adds a metric for the value in question (size of in-memory remap
trace) that will help us prove or disprove the theory if the issue
happens again in the future.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>