catalog: Move stash upgrade code path to catalog
This commit moves the code responsible for detecting and executing
catalog data migrations from the stash crate to the catalog crate. This
code contained logic that was specific to the catalog create.
Specifically, it required a collection called "config" to exist, and it
required a key called "user_version" in that collection. Previously,
this logic was only usable by the catalog, so it makes a bit more sense
to move that code to the catalog. Also, it will help us generalize
catalog data migrations to work with the persist catalog.

Works towards resolving #22977