storage: fix stuck txns shard since caused by long-lived TxnsCache
We had this TxnsCache in the storage controller to serve snapshot
queries, but there isn't an easy way to ensure it continually makes
progress, which results in the txns shard since getting stuck.

We only call this method at boot, so instead create short-lived ones.
It's wasteful to create this TxnsCache and then throw it away for each
of them, but it's better than the alternative. Replace this with the
shared TxnsCache thing we'll have to do anyway for the dataflow
operators.