adapter: improve GlobalMirPlan until resolution
Previously this would have triggered the as_of overflow warning:

```
create view v as select 1;
select * from v;
```

Because that's a query not on a timeline, so we use the MAX timestamp
which overflows when incremented.

Instead, detect if we are in a timeline with no timestamp and don't even
try to set the until.