npm: do not tag stable builds as latest if newer versions are published.
We currently tag the newest stable version of our WASM packages as `latest`,
even if the version on npm is greater. This is seen when we create a patch
release for the currently-deployed version of Materialize after we cut
the next release candidate.

The `latest` tag allows tooling, such as dependabot, to track the latest
stable version.