persist-txn: correctly handle restarts flipping the feature on and off
There are a bunch of tricky issues that result from the possibility of
two concurrently running environmentd instances with different settings
for whether persist-txn tables or the legacy tables are used. This
installs the necessary fencing to prevent incorrectness from happening.
See the big comment introduced in the commit for how it works.