tests: use new user timeline per test and testdrive invocation
Before, it could happen that tests that thought they had sole control
over the read/write timestamp of a timeline would interact at a
distance. For example, a test might create a CDCv2 source and advance
it's upper to `2`, but another test might already have advanced that
shared timeline's read/write timetamp to `3`, meaning queries that are
expected to return in the first test would block forever.

User timelines are not available in production, so I'm hesitant to put
in more work than this quick fix.

Fixes #23096