stash: Remove storage proto objects from stash (#23009)
The current users of the stash are the adapter for the catalog and the
storage-controller for metadata. All objects that users of the stash
store in the stash, are defined within the mz-stash-types crate. All
objects across both uses are stored in a file called objects.proto, and
a hash of objects.proto is compiled into the binary. If anyone changes
anything in objects.proto then Materialize fails to compile unless an
explicit migration is written for that object using the stash migration
framework.

However, if any object is an object that was used by the storage
controller is changed, then the data within the stash is never actually
migrated. Even if a migration is written, the storage data will not be
migrated. The reason is that the stash user needs to create a
collection named "config" and store a value with the key
"user_version", and then manually call the upgrade function in the
Stash.

This commit moves all the storage protobuf objects out of the
mz-stash-types crate and into the mz-storage-types crates. In addition
to avoiding issues around migrating storage objects, this is a step in
the direction of moving the migration responsibility from the stash to
the catalog.

Works towards resolving #22977