mzcompose: use name for Materialized service hostname
By default, docker compose uses the container's ID for its hostname,
however that isn't stable and makes inconsistent any state that
depends on the hostname. In our experience, this means that the
persist location's URI is unstable, which causes some of our internal
asserts to fire.

This commit attempts to stabilize the hostname by making it match
the service's name.