persist-txn: fix bug in storage controller snapshot{,_cursor}
These methods were using `self.txns_id` as a switch for whether to do
the read through txns or directly, but this is not correct. This field
being Some only indicates that txns are enabled, we have to actually
look at the collection metadata's `txns_shard` field to see if this
particular collection is using txns or not.

In practice, this meant that if we had the txns feature flag on, then
these two methods were using it for all reads (including sources, etc)
not just tables. `snapshot` happens to only be called on tables, so it
was wrong but happened to not be causing any issues in practice.
`snapshot_cursor`, however, is used on all sorts of things when the
feature flag is on (on in CI, but not on yet in staging or prod).

Cause by CI with the enable_persist_txn_tables flag on, so we don't need
a regression test.