test: rewrite Kafka auth tests
The kafka-ssl and kafka-sasl-plain compositions were complicated enough
that they couldn't be easily extended to support other authentication
configurations (e.g., SASL_PLAINTEXT).

This commit merges the existing compositions into a new kafka-auth
composition that is capable of exercising every possible Kafka
authentication configuration that we support.

It also adjusts the default connection validation timeout from 30s to
10s. Waiting 30s to see error messages in tests is painful, and likely
reflective of real user pain when creating connections in production.  A
10s timeout is long enough for any remotely healthy external system to
respond to validation, and is short enough that it's not painful to wait
out the timeout.