Upgrade Python deps to be 3.12-compatible.
Python 3.12 is stable, so lets bump some of our deps to ensure
Materialize can be developed with that version. Upgrades include:

1. pip + setuptools to deal with removal of the long-depracated
  ImpImporter.
2. ruff + psycopg + scipy to the first version to support 3.12 wheels

ruff, psycopg, and scipy aren't being updated to the newest versions in
an attempt to minimize the diff introduced by this PR. ruff, for
example, is at 0.1.5 at the moment. We can upgrade it and others in
subsequent PRs.