catalog: Move catalog config (#23120)
This PR moves the adapter::catalog::config module to the catalog
crate, by removing the dependency on the adapter::config module from
the adapter::catalog::config module. This probably creates an
unnecessary level of indirection, but is useful for moving the
adapter::catalog::config module to the catalog crate. At some point in
the future we will actually need remove the adapter::config module as a
dependency from the entire adapter::catalog module. However, we can
cross that bridge when we get there.

Works towards resolving # #22593