test: Disable shadow catalog in tests which kill materialized
The Shadow catalog does the following on every catalog operation:

    1. Execute the operation on catalog A.
    2. Execute the operation on catalog B.
    3. Compare the results.

The parallel-workload tests will randomly kill Materialize, so we could
end up in a situation where the operation happens on catalog A but not
on catalog B, and we will see an inconsistency between the two
catalogs. The shadow catalog is meant to be used to compare two catalog
implementations, so it's OK that it has this deficiency. Given that,
this commit disables the shadow catalog in parallel-workload tests.

Fixes #23146