storage: fix bug in SinkAsOf::downgrade()
Before, when our frontier matched the since, we would still flip the
`strict` bit, while we should instead leave the `SinkAsOf` completely
alone.

This was surfacing as part of the change to initialize the sink as_of to
the `least_valid_upper` instead of the timeline's current `read_ts`.
That change makes it so we read some sources at `as_of = [0]`, meaning
`SinkAsOf{ frontier: [0], strict: false }`, and that gets flipped (by
the downgrade logic) to `SinkAsOf{ frontier: [0], strict: true }`, and
then no data (meaning the snapshot, really) makes it to the sinks.

We need to _carefully_ look at this to make sure we don't introduce a
data duplication bug in the sink. Dealings right around the `since` are
always tricky.