storage: upon sink creation, pick lowest valid as_of instead of oracle read_ts
Some test (kafka-avro-upsert-sinks.td) assume that a sink is created
with an as_of of [0] when it is created right after the sunk source,
when that source is a CDCv2 source whose upper is held at 0. With an
as_of of 0, we sink data with "full fidelity", meaning records that have
different timestamps on the input show up with different timestamps (in
different transactions) on the sinks, which these kinds of tests make
use of.

There is a complicated interaction of this with a recent (currently
reverted) PR (https://github.com/MaterializeInc/materialize/pull/23022)
that made it so we never remove timeline state:
 - Before this change, we would pick the source timeline's read_ts as
   the as_of. When the timeline hadn't existed before (or when it had
   existed, then was deleted, and then was re-created again), this would
   be 0.
 - With the change to timeline state there could be leftover timeline
   state from other tests, or from a previous invocation of the same
   test. We would therefore now pick some non-zero timestamp when
   creating the sink, meaning we wouldn't get the expected fidelity in
   terms of timestamps.

This change makes it so we create sinks at the lowest allowed as_of,
side-stepping the timestamp oracle, and therefore side-stepping the
complicated interaction.

Incidentally, this change makes it so the sink as_of is initialized the
way materialized views already do it:
https://github.com/MaterializeInc/materialize/blob/9ca25803650d36018fd5ad96b2b296e6e16b959e/src/adapter/src/coord/sequencer/inner.rs#L1042