Implement HeapSize in rendering
This commit provides implementations for the HeapSize trait in
rendering. These implementations improve the memory attribution of data
stored in arrangements. In particular, the attribution of tuples of
(time, diff) is significantly improved by encoding size estimates in a
manner that is sensitive to types used for this purpose. One limitation
of the present implementation is that estimation for an entire
Vec<(time, diff)> is achieved by sampling one of the elements of the
vector. This will result in precise estimates for fixed-length tuples
as in non-recursive accumulable reductions; however, where lengths are
variable, as in hierarchical reductions or top-k, the strategy may
lead to inaccuracies.