adapter: Update object dependency tracking
Previously, Materialize had two notions of object dependencies.

  1. An object required to determine the timestamp of a
  `MirRelationExpr` in a plan.
  2. An object directly referenced in SQL. This is represented in the
  code as `ResolvedIds`

However, there's actually a third type of dependency:

  3. An object required while planning an
  `mz_sql_parser::ast::Statement`.

This new group contains all the objects in group (2), but may also
contain additional objects needed for planning. For example, certain
functions and casts are re-written as SQL statements that includes
catalog objects. Also, `SHOW` commands are re-written as SQL that
includes catalog objects. When determining the order to create objects,
we must use group (3), not (2).

This commit adds the concept of group (3) into the code and uses it
where applicable.

Resolves #23153