catalog: Fix shadow timestamp discrepancies (#23219)
The shadow catalog is a catalog implementation used for testing.
Internally it has a stash and persist catalog implementation. On every
operation it compares the results of each implementation and panics if
there's a difference.

The Coordinator will update the timestamps of every timeline on a set
interval, currently set to 10 seconds. Timestamps are currently stored
in the catalog. If the Coordinator is shut down during this interval,
then it's possible that only one of the catalog implementations updated
the timestamps, while the other didn't. This makes it impossible to
correctly test scenarios involving restarts with the shadow catalog.

This commit fixes this issue, by removing any discrepancies in the
timestamps across the stash and persist when opening the catalog. The
shadow catalog will treat the larger of the two timestamps as correct
and update accordingly.