storage: refactor sinks
This is a refactor and simplification of sinks. My primary motivation
that led to this PR were the numerous non-standard frontier operations
that we had in the code which had to be very carefully choreographed
when they interacted with the rest of the system. Some examples of that
were:

* Having a special form of as_of frontiers that contained a `strict`
  flag which was flipped during downgrades
* Storing the maximum timestamp of a batch in the progress topic instead
  of an upper frontier
* Reporting the upper - 1 back to the controller for.. reasons
* Maybe others that I forgot

This commit removes all these special cases and makes sinks behave as much
as possible as a normal materialized view. Pristine frontiers are
committed and reported.

Beyond these frontier simplification this PR also:
* Integrates the fallible async operator paradigm with the sink
  infrastructure so that the errors can be handled in a more natural way
* Brings back parallel encoding of values before writing them to kafka