adapter: Box several very large `Future`s (#23226)
This PR updates a few functions in the adapter to return Box-ed Futures
instead of a Future that is put on the stack, i.e. `async fn` gets
changed to `fn ... -> BoxFuture`.

### Motivation

While working on improving our environmentd Rust tests, e.g.
https://github.com/MaterializeInc/materialize/issues/22432, I ran into a
few stack overflows. It's known that the [coordinator stores a lot of
state on its
stack](https://github.com/MaterializeInc/materialize/blob/main/src/adapter/src/coord.rs#L2094-L2097),
but it wasn't the coordinator thread that was panicking, it was the
thread created by the test framework to run the test case. Increasing
our `STACK_RED_ZONE` size allowed the tests to pass, but that didn't
feel like a good solution.

Using `RUSTFLAGS=-Zprint-type-sizes` I was able to see the sizes for all
compiled types, and I found that we have several Futures returned by
`async fn`s that are 10s of KB in size! The Coordinator's main `async fn
serve` returned a Future that was `92KB`! Based on the sizes reported
and some hot loops in the code base, I replaced a few `async fn` with
`fn -> BoxFuture`. Even though these Futures are now stored on the heap
and not the stack, there is a good chance we could see a performance
improvement from this change, because previously these very large
Futures could get memcopy-d when we run them, but now they'd stay on the
heap and we'd be passing around a pointer.

Ideally we'd have a proc macro like `#[box_future]` that would do this
boxing automatically and then we can enable the
[large_futures](https://rust-lang.github.io/rust-clippy/master/#/large_futures)
clippy lint, but that's out of scope for this PR.

### Tips for reviewer

The diff is **much** smaller if viewed with whitespace hidden.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a