persist-txn: introduce a sharable TxnsCache called TxnsRead
Each `txns_progress` operator currently gets a unique TxnsCache and each
of these have a subscription to the txns shard. This causes a number of
issues:

- It's wasteful of cpu, memory, and network.
- It's easy to have stability issues if the since of the txns_shard
  falls behind at all. There are regular writes to this from the metrics
  that we keep in system tables. If the since falls behind, we don't get
  to consolidate the retractions out and this history ends up being read
  `num_worker` times each time a `persist_source` operator is rendered.
  Even in a fresh, empty environment, the introspection cluster has
  quite a number of these by default. Before I worked out the last of
  the since hold bugs, my staging env would be using half a core
  continuously just on these subscriptions. This is too brittle.
- A subscription requires registering a persist reader, which is
  linearized by the shard. So a query, including transient ones, with
  `I` inputs and `W` workers has to linearize `I*W` registrations. This
  is causing timeouts in nightly CI in my branch to enable persist-txn
  by default.

The last is the immediate motivation for this PR. The fix is to have a
single TxnsCache process-wide and share it with all the txns_progress
operators.

This commit establishes the skeleton of a new shared TxnsCache (called
TxnsRead) which runs in a task and is communicated with over a channel,
like the "persist workers" in the storage controller. It intentionally
punts on a number of the hard problems (compaction, cancellation, etc)
to break up the code review a bit. Plumbing it process-wide will be
quite a large diff, so that will also be a follow up PR (this just hooks
it up 1:1 inside txns_progress so that we can get test coverage via
existing tests).