Enable jemalloc profiling by default
This commit enables jemalloc profiling by default. It is already enabled
everywhere in staging and prod via feature flag, but the part of process
startup that occurs before the process gets access to the dynamic
configuration value still has memory profiling disabled due to the
default. This may cause us to miss memory events that happen early in
process startup, so we want profiling to be enabled by default.