storage: fix yet another shutdown bug
As part of #20150 I changed the decode operator to use the
`AsyncOperatorBuilder::build_fallible` construction so that we could use
the ergonomic `?` operator and trigger safe dataflow restarts when
transient errors were hit.

The way `build_fallible` makes using `?` shutdown-safe is by not giving
you access to owned `Capability` values like the normal builder.
Instead, it gives out mutable references to capabilities which can be
used for all the normal things (sending data/downgrades) but when a `?`
triggers an error and the future returns we only drop a reference, which
doesn't have any side effects. The actual capability is kept alive by
the `build_fallible` machinery until dataflow shutdown has been
initiated by all workers.

Unfortunately that pattern completely breaks down with operators that
don't want to keep around any initial capabilities and instead just want
to react on input capabilities, which is exactly what the decode
operator wants to do. Traditionally, operators that just transform their
input in some way want to immediately drop their initial capabilities
and produce data to their outputs using the input capabilities received.
In the fallible operator land this is not shutdown-safe! This is because
the capabilities received by calling `input_handle.next_mut()` are
normal, owned `InputCapability` values, and if an error is hit which
causes the return of the operator's `Future` then that capability will
be dropped without waiting for all workers to initiate shutdown.

I have verified that this can happen and can lead to data loss. I
created a source that read a single avro record from kafka but when it
tried to decode it it hit a schema registry error (because I had stopped
the CSR container). This led to the decode operator to drop the
`InputCapability` of that message as part of erroring out, which caused
the frontier to advance, which caused the persist sink to happily
advance the source upper with no data. When the dataflow eventually
restarted the resume upper had moved past the offset of the original
message and that message was permanently lost.

This PR fixes the issue by keeping around the initial capabilities and
driving them around every time the input frontier advances.

PS: This is yet another subtle bug on the big pile of shutdown bugs that
we've had to deal with. Both this issue and all the previous ones would
not have manifested if we had been using `Worker::drop_dataflow` to drop
dataflows.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>