catalog: Fix read-only shadow ts discrepancies (#23229)
The shadow catalog is a catalog implementation used for testing.
Internally it has a stash and persist catalog implementation. On every
operation it compares the results of each implementation and panics if
there's a difference.

The Coordinator will update the timestamps of every timeline on a set
interval, currently set to 10 seconds. Timestamps are currently stored
in the catalog. If the Coordinator is shut down during this interval,
then it's possible that only one of the catalog implementations updated
the timestamps, while the other didn't. Writeable catalogs will fix any
timestamp discrepancies during startup. However, read-only catalogs
cannot fix this discrepancy because they cannot write.

This commit fixes this issue by having the shadow catalog ignore any
timestamp discrepancies during read-only mode.