catalog: Remove read handle from persist catalog
The persist catalog only needs a read handle during bootstrapping, but
it was holding on to one for the entire lifetime of the catalog. This
commit removes the persist read handle from the persist catalog
implementation. Instead, the persist catalog will only open a read
handle during bootstrapping when it needs one.