Don't advance nonzero timestamps in the Postgres client
Outside of a snapshot scenario, it's bad for the slot's LSN to be ahead
of where we think it should be... that would mean we've missed data
that's not reflected in our internal state. In our brave new world of
restorable backups, this is now more likely! It sounds like it might not
be safe to just never fast-forward the offsets of individual subsources,
but at least we should only do that to the subsources that really need
it.