rbac: Allow mz_support to read progress sources
This commit updates the default privileges such that mz_support can
read all progress sources. The default privilege framework is not
granular enough to create default privileges for specific types of
sources, so this default privilege is hardcoded into the binary and
not visible via mz_default_privileges.

This commit also adds a migration to grant SELECT on all existing
progress sources to mz_support.

Resolves #23199