storage: speed up sink creation when `DescribeConfigs` is banned
Creating a Kafka sink was previously slow when the Kafka cluster
prohibited Materialize from using the `DescribeConfigs` operation. There
were two problems:

 1. The code to check whether the progress topic referenced a deleted
    data topic was racing with the code to create the progress topic.
    The reference check would win the race, then choke on the fact that
    the progress topic did not yet exist. Materialize would suspend and
    resume the sink after 30s, at which point the progress topic would
    already exist and the reference check would succeed, but then...

 2. Materialize's topic creation code path would attempt to create the
    sink's data topic with -1 partitions, indicating that the broker's
    default number of partitions should be used, but then Materialize
    would *wait* for the topic to have -1 partitions, which of course
    never happened. Topic creation would eventually time out, at which
    point Materialize would suspend and resume the sink, at which point
    the data topic would already exist and so Materialize wouldn't wait
    for it to have -1 partitions.

This commit fixes (1) by teaching the reference check to handle missing
progress topics and (2) by teaching the topic creation code not to wait
for the topic to have any particular number of partitions if the number
of requested partitions was -1. It also adds a test that ensures Kafka
sinks are created without ever entering the `stalled` status.